move entities to controlled folder
