+++ .gitignore
@@ -1,2 +1,10 @@
-target/*
+# Visual Studio Code #
 .vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+!.vscode/*.code-snippets
+
+target/*
+
