+++ pom.xml
@@ -77,6 +77,7 @@
   </dependencies>
 
 
+  <!-- 
   <build>
     <plugins>
       <plugin>
@@ -101,6 +102,6 @@
       </plugin>
     </plugins>
   </build>
-
+-->
 
 </project>
