+++ src/main/java/hust/kien/project/dao/authordao/AuthorLibraryDao.java
@@ -1,7 +1,7 @@
 package hust.kien.project.dao.authordao;
 
 import hust.kien.project.dao.LibraryDao;
-import hust.kien.project.model.Author;
+import hust.kien.project.model.author.Author;
 
 public interface AuthorLibraryDao extends LibraryDao<Author, Long> {
