+++ src/main/java/hust/kien/project/model/book/BookGenre.java
@@ -1,4 +1,4 @@
-package hust.kien.project.model;
+package hust.kien.project.model.book;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
